requirements.txt
selenium
undetected-chromedriver
python-dotenv
setuptools
webdriver-manager